chore: ignore expired txs for approval state
