+++ apps/cowswap-frontend/src/legacy/hooks/useRecentActivity.ts
@@ -6,12 +6,20 @@ import { SupportedChainId, SupportedChainId as ChainId } from '@cowprotocol/cow-
 import { UiOrderType } from '@cowprotocol/types'
 import { useWalletInfo } from '@cowprotocol/wallet'
 
-import ms from 'ms.macro'
-
-import { isTransactionRecent, useAllTransactions, useTransactionsByHash } from 'legacy/state/enhancedTransactions/hooks'
+import { useAllTransactions, useTransactionsByHash } from 'legacy/state/enhancedTransactions/hooks'
 import { EnhancedTransactionDetails } from 'legacy/state/enhancedTransactions/reducer'
 import { Order, OrderStatus } from 'legacy/state/orders/actions'
 import { useCombinedPendingOrders, useOrder, useOrders, useOrdersById } from 'legacy/state/orders/hooks'
+
+import { getIsTxExpired } from 'modules/onchainTransactions/utils/getIsTxExpired'
+
+/**
+ * Returns whether a transaction happened in the last day (86400 seconds * 1000 milliseconds / second)
+ * @param tx to check for recency
+ */
+function isTransactionRecent(tx: EnhancedTransactionDetails): boolean {
+  return new Date().getTime() - tx.addedTime < 86_400_000
+}
 
 export interface AddedOrder extends Order {
   addedTime: number
@@ -42,16 +50,6 @@ export enum ActivityStatus {
 enum TxReceiptStatus {
   PENDING,
   CONFIRMED,
-}
-
-const ONCHAIN_TX_TIMEOUT: Record<SupportedChainId, number> = {
-  [SupportedChainId.MAINNET]: ms`30m`,
-  [SupportedChainId.GNOSIS_CHAIN]: ms`15m`,
-  [SupportedChainId.SEPOLIA]: ms`5m`,
-}
-
-function isTxExpired(tx: EnhancedTransactionDetails, chainId: SupportedChainId): boolean {
-  return Date.now() - tx.addedTime > ONCHAIN_TX_TIMEOUT[chainId]
 }
 
 /**
@@ -106,7 +104,7 @@ export function useRecentActivity(): TransactionAndOrder[] {
             id: tx.hash,
             status: tx.receipt
               ? OrderStatus.FULFILLED
-              : isTxExpired(tx, chainId)
+              : getIsTxExpired(tx, chainId)
               ? OrderStatus.EXPIRED
               : OrderStatus.PENDING,
           }
@@ -187,7 +185,7 @@ export function createActivityDescriptor(
     const isReceiptConfirmed =
       !!tx.receipt && (tx.receipt.status === TxReceiptStatus.CONFIRMED || typeof tx.receipt.status === 'undefined')
     const isCancelTx = tx?.replacementType === 'cancel'
-    isPending = !tx.receipt && !isTxExpired(tx, chainId)
+    isPending = !tx.receipt && !getIsTxExpired(tx, chainId)
     isPresignaturePending = false
     isConfirmed = !isPending && isReceiptConfirmed
     isCancelling = isCancelTx && isPending
